pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "llm-rag"
version = "0.1.0"
description = "Bilingual (FR/EN) RAG system skeleton with cloud LLM + local Ollama embeddings/fallback."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Jean", email = "" }
]
keywords = ["rag", "ollama", "llm", "retrieval", "embeddings"]

dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.0",
  "requests>=2.31.0",
  "chromadb>=0.4.22",
  "pypdf>=4.0.0",
  "beautifulsoup4>=4.12.0",
  "markdown>=3.5.0",
  "pyyaml>=6.0.1",
  "python-dotenv>=1.0.0",
  "tqdm>=4.66.0"
]

[project.optional-dependencies]
local-dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.1.0",
  "ruff>=0.4.0",
  "mypy>=1.8.0",
  "types-requests>=2.31.0.20240218"
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/rag"]
